Using GAT in order to simplify templating
GAT is used to represent a RenderScene in a given rendering system (Wgpu or Webgl).
The concept of RenderScene is a GAT on the RenderTarget trait. Since everything is triangles right now, the TriangleScene trait became the RenderScene trait.
When we have different kind of drawing primitives, we will flexibilize more the concept of RenderScene.